chore: cache refactor
