+++ layer/providers/cacheApi/spot.ts
@@ -1,21 +1,40 @@
 import { SpotMarket, AllChronosSpotMarketSummary } from '@injectivelabs/sdk-ts'
 import { BaseCacheApi } from './base'
 import { indexerRestSpotChronosApi, indexerSpotApi } from '../../Service'
+import { IS_MAINNET } from '~/utils/constant'
 
 export class SpotCacheApi extends BaseCacheApi {
   async fetchMarkets() {
+    const fetchFromExchange = async () => {
+      const response = await indexerSpotApi.fetchMarkets()
+
+      return response
+    }
+
+    if (!IS_MAINNET) {
+      return fetchFromExchange()
+    }
+
     try {
       const response = await this.client.get<SpotMarket[]>('/spot/markets')
 
       return response.data
     } catch (e) {
-      const response = await indexerSpotApi.fetchMarkets()
-
-      return response
+      return fetchFromExchange()
     }
   }
 
   async fetchMarketsSummary() {
+    const fetchFromExchange = async () => {
+      const response = await indexerRestSpotChronosApi.fetchMarketsSummary()
+
+      return response
+    }
+
+    if (!IS_MAINNET) {
+      return fetchFromExchange()
+    }
+
     try {
       const response = await this.client.get<AllChronosSpotMarketSummary[]>(
         '/spot/summary'
@@ -23,9 +42,7 @@ export class SpotCacheApi extends BaseCacheApi {
 
       return response.data
     } catch (e) {
-      const response = await indexerRestSpotChronosApi.fetchMarketsSummary()
-
-      return response
+      return fetchFromExchange()
     }
   }
 }
