+++ layer/providers/cacheApi/token.ts
@@ -1,9 +1,22 @@
 import { Pagination, TotalSupply } from '@injectivelabs/sdk-ts'
 import { BaseCacheApi } from './base'
 import { bankApi } from '../../Service'
+import { IS_MAINNET } from '~/utils/constant'
 
 export class TokenCacheApi extends BaseCacheApi {
   async fetchTotalSupply() {
+    const fetchFromBank = async () => {
+      const { supply, pagination } = await bankApi.fetchTotalSupply({
+        limit: 2000
+      })
+
+      return { supply, pagination }
+    }
+
+    if (!IS_MAINNET) {
+      return fetchFromBank()
+    }
+
     try {
       const response = await this.client.get<{
         supply: TotalSupply
@@ -12,11 +25,7 @@ export class TokenCacheApi extends BaseCacheApi {
 
       return response.data
     } catch (e) {
-      const { supply, pagination } = await bankApi.fetchTotalSupply({
-        limit: 2000
-      })
-
-      return { supply, pagination }
+      return fetchFromBank()
     }
   }
 }
